Shape Expressions 1.0 Definition W3C Member Submission 2 June 2014
This is the official member submission. It contains multiple serious
errors in the translation of the Z Notation from LaTeX to HTML. These
need to be corrected and the version on W3C should be updated. The Z
Notation has several stylistic issues and semantic errors which also
should be corrected.